% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_tweet.R
\name{include_tweet}
\alias{include_tweet}
\title{Include A Tweet in All R Markdown Formats}
\usage{
include_tweet(tweet_url, plain = FALSE, ...)
}
\arguments{
\item{tweet_url}{The URL for the tweet, or use \code{\link[=tweet_url]{tweet_url()}} to build the URL
from a \code{screen_name} and \code{status_id}.}

\item{plain}{Embeds the tweet without including Twitter's JavaScript code.
This means the tweets will be displayed as minimally styled HTML in
\code{\link[=tweet_embed]{tweet_embed()}} or \code{\link[=tweet_screenshot]{tweet_screenshot()}}. In \code{\link[=include_tweet]{include_tweet()}},
\code{plain = TRUE} will convert the tweet text into markdown for non-HTML
outputs.

Note that for HTML outputs, you can't selectively choose whether or not a
tweet is rendered in plain HTML if you include multiple tweet. Including
one rich embedded tweet will trigger rich embedding of all tweets loaded
on the page.}

\item{...}{
  Arguments passed on to \code{\link[=tweet_embed]{tweet_embed}}
  \describe{
    \item{\code{maxwidth}}{The maximum width of a rendered Tweet in whole pixels. A
supplied value under or over the allowed range will be returned as the
minimum or maximum supported width respectively; the reset width value will
be reflected in the returned \code{width} property. Note that Twitter does not
support the oEmbed \code{maxheight} parameter. Tweets are fundamentally text,
and are therefore of unpredictable height that cannot be scaled like an
image or video. Relatedly, the oEmbed response will not provide a value for
\code{height.} Implementations that need consistent heights for Tweets should
refer to the \code{hide_thread} and `hide_media`` parameters below.}
    \item{\code{hide_media}}{When set to \code{true}, \code{"t"}, or \code{1} links in a Tweet are not
expanded to photo, video, or link previews.}
    \item{\code{hide_thread}}{When set to \code{true}, \code{"t"}, or \code{1} a collapsed version of
the previous Tweet in a conversation thread will not be displayed when the
requested Tweet is in reply to another Tweet.}
    \item{\code{omit_script}}{When set to \code{FALSE}, Twitter's remote script responsible
for loading the Tweet widget will be used. Otherwise, the script included
by this package will be used. If you are including multiple tweets on a
page, it is recommended to use the default \code{FALSE}.}
    \item{\code{align}}{Specifies whether the embedded Tweet should be floated left,
right, or center in the page relative to the parent element.}
    \item{\code{related}}{A comma-separated list of Twitter usernames related to your
content. This value will be forwarded to
\href{https://developer.twitter.com/web/intents}{Tweet action intents}
if a viewer chooses to reply, like, or retweet the embedded Tweet.}
    \item{\code{lang}}{Request returned HTML and a rendered Tweet in the specified
\href{https://developer.twitter.com/web/overview/languages}{Twitter language supported by embedded Tweets}.}
    \item{\code{theme}}{When set to \code{dark}, the Tweet is displayed with light text over
a dark background.}
    \item{\code{link_color}}{Adjust the color of Tweet text links with a
\href{https://en.wikipedia.org/wiki/Web_colors#Hex_triplet}{hexadecimal color value}.}
    \item{\code{widget_type}}{Set to \code{video} to return a Twitter Video embed for the
given Tweet.}
    \item{\code{dnt}}{When set to \code{true}, the Tweet and its embedded page on your site
are not used for purposes that include
\href{https://support.twitter.com/articles/20169421}{personalized suggestions}
and \href{https://support.twitter.com/articles/20170405}{personalized ads}.}
  }}
}
\value{
An \code{htmltools::tagList()} to include a tweet in an HTML document, or
a screen shot of the tweet for use in non-HTML outputs.
}
\description{
Similar to \code{\link[knitr:include_graphics]{knitr::include_graphics()}}, but for tweets. In HTML documents,
the tweet is embedded using \code{\link[=tweet_embed]{tweet_embed()}} and for all other documents types
a screen shot of the tweet is rendered and used \code{\link[=tweet_screenshot]{tweet_screenshot()}}. If you
would rather that just the text of the tweet be included in non-HTML outputs,
use \code{\link[=tweet_embed]{tweet_embed()}}.
}
\examples{

include_tweet("https://twitter.com/dsquintana/status/1275705042385940480")

}
\seealso{
Other Tweet-embedding functions: 
\code{\link{tweet_embed}()},
\code{\link{tweet_screenshot}()}
}
\concept{Tweet-embedding functions}
