% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{tweet_embed}
\alias{tweet_embed}
\alias{tweet_url}
\alias{\%tweet\%}
\title{Embed a Tweet in R Markdown HTML outputs}
\usage{
tweet_embed(
  tweet_url,
  maxwidth = 550,
  plain = FALSE,
  hide_media = FALSE,
  hide_thread = FALSE,
  omit_script = TRUE,
  align = c("none", "left", "right", "center"),
  related = NULL,
  lang = "en",
  theme = c("light", "dark"),
  link_color = NULL,
  widget_type = NULL,
  dnt = TRUE
)

tweet_url(screen_name, status_id)

screen_name \%tweet\% status_id
}
\arguments{
\item{tweet_url}{The URL for the tweet, or use \code{\link[=tweet_url]{tweet_url()}} to build the URL
from a \code{screen_name} and \code{status_id}.}

\item{maxwidth}{The maximum width of a rendered Tweet in whole pixels. A
supplied value under or over the allowed range will be returned as the
minimum or maximum supported width respectively; the reset width value will
be reflected in the returned \code{width} property. Note that Twitter does not
support the oEmbed \code{maxheight} parameter. Tweets are fundamentally text,
and are therefore of unpredictable height that cannot be scaled like an
image or video. Relatedly, the oEmbed response will not provide a value for
\code{height.} Implementations that need consistent heights for Tweets should
refer to the \code{hide_thread} and `hide_media`` parameters below.}

\item{plain}{Embeds the tweet without including Twitter's JavaScript code.
This means the tweets will be displayed as minimally styled HTML in
\code{\link[=tweet_embed]{tweet_embed()}} or \code{\link[=tweet_screenshot]{tweet_screenshot()}}. In \code{\link[=include_tweet]{include_tweet()}},
\code{plain = TRUE} will convert the tweet text into markdown for non-HTML
outputs.

Note that for HTML outputs, you can't selectively choose whether or not a
tweet is rendered in plain HTML if you include multiple tweet. Including
one rich embedded tweet will trigger rich embedding of all tweets loaded
on the page.}

\item{hide_media}{When set to \code{true}, \code{"t"}, or \code{1} links in a Tweet are not
expanded to photo, video, or link previews.}

\item{hide_thread}{When set to \code{true}, \code{"t"}, or \code{1} a collapsed version of
the previous Tweet in a conversation thread will not be displayed when the
requested Tweet is in reply to another Tweet.}

\item{omit_script}{When set to \code{FALSE}, Twitter's remote script responsible
for loading the Tweet widget will be used. Otherwise, the script included
by this package will be used. If you are including multiple tweets on a
page, it is recommended to use the default \code{FALSE}.}

\item{align}{Specifies whether the embedded Tweet should be floated left,
right, or center in the page relative to the parent element.}

\item{related}{A comma-separated list of Twitter usernames related to your
content. This value will be forwarded to
\href{https://developer.twitter.com/web/intents}{Tweet action intents}
if a viewer chooses to reply, like, or retweet the embedded Tweet.}

\item{lang}{Request returned HTML and a rendered Tweet in the specified
\href{https://developer.twitter.com/web/overview/languages}{Twitter language supported by embedded Tweets}.}

\item{theme}{When set to \code{dark}, the Tweet is displayed with light text over
a dark background.}

\item{link_color}{Adjust the color of Tweet text links with a
\href{https://en.wikipedia.org/wiki/Web_colors#Hex_triplet}{hexadecimal color value}.}

\item{widget_type}{Set to \code{video} to return a Twitter Video embed for the
given Tweet.}

\item{dnt}{When set to \code{true}, the Tweet and its embedded page on your site
are not used for purposes that include
\href{https://support.twitter.com/articles/20169421}{personalized suggestions}
and \href{https://support.twitter.com/articles/20170405}{personalized ads}.}

\item{screen_name}{The user's screen name}

\item{status_id}{The tweet's status id}
}
\value{
The placeholder HTML for the tweet
}
\description{
Uses Twitter's \href{https://publish.twitter.com}{oembed} API to embed a
tweet in R Markdown HTML outputs.
}
\section{Functions}{
\itemize{
\item \code{tweet_url}: Create a URL for a tweet from a screen name and status id.

\item \code{\%tweet\%}: A shortcut for the impatient.
}}

\examples{
\dontrun{
# these are equivalent
tweet_embed("https://twitter.com/alexpghayes/status/1211748406730706944")

tweet_embed(tweet_url("alexpghayes", "1211748406730706944"))

# Adding many tweets at once to a document takes a little bit more work
thread <- c(
  "https://twitter.com/grrrck/status/1333804309272621060",
  "https://twitter.com/grrrck/status/1333804487148855300",
  "https://twitter.com/grrrck/status/1333805092152123394"
)

htmltools::tagList(
  lapply(thread, tweet_embed, plain = TRUE)
)
}

}
\references{
\url{https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-oembed}
}
\seealso{
Other Tweet-embedding functions: 
\code{\link{include_tweet}()},
\code{\link{tweet_screenshot}()}
}
\concept{Tweet-embedding functions}
