% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{tweet_screenshot}
\alias{tweet_screenshot}
\title{Take or embed a screenshot of a tweet}
\usage{
tweet_screenshot(tweet_url, maxwidth = 550, file = NULL, ...)
}
\arguments{
\item{tweet_url}{The URL for the tweet, or use \code{\link[=tweet_url]{tweet_url()}} to build the URL
from a \code{screen_name} and \code{status_id}.}

\item{maxwidth}{The maximum width of a rendered Tweet in whole pixels. A
supplied value under or over the allowed range will be returned as the
minimum or maximum supported width respectively; the reset width value will
be reflected in the returned \code{width} property. Note that Twitter does not
support the oEmbed \code{maxheight} parameter. Tweets are fundamentally text,
and are therefore of unpredictable height that cannot be scaled like an
image or video. Relatedly, the oEmbed response will not provide a value for
\code{height.} Implementations that need consistent heights for Tweets should
refer to the \code{hide_thread} and `hide_media`` parameters below.}

\item{file}{A character string of output files. Can end with \code{.png} or \code{.pdf}.}

\item{...}{
  Arguments passed on to \code{\link[=tweet_embed]{tweet_embed}}
  \describe{
    \item{\code{hide_media}}{When set to \code{true}, \code{"t"}, or \code{1} links in a Tweet are not
expanded to photo, video, or link previews.}
    \item{\code{hide_thread}}{When set to \code{true}, \code{"t"}, or \code{1} a collapsed version of
the previous Tweet in a conversation thread will not be displayed when the
requested Tweet is in reply to another Tweet.}
    \item{\code{omit_script}}{When set to \code{FALSE}, Twitter's remote script responsible
for loading the Tweet widget will be used. Otherwise, the script included
by this package will be used. If you are including multiple tweets on a
page, it is recommended to use the default \code{FALSE}.}
    \item{\code{align}}{Specifies whether the embedded Tweet should be floated left,
right, or center in the page relative to the parent element.}
    \item{\code{related}}{A comma-separated list of Twitter usernames related to your
content. This value will be forwarded to
\href{https://developer.twitter.com/web/intents}{Tweet action intents}
if a viewer chooses to reply, like, or retweet the embedded Tweet.}
    \item{\code{lang}}{Request returned HTML and a rendered Tweet in the specified
\href{https://developer.twitter.com/web/overview/languages}{Twitter language supported by embedded Tweets}.}
    \item{\code{theme}}{When set to \code{dark}, the Tweet is displayed with light text over
a dark background.}
    \item{\code{link_color}}{Adjust the color of Tweet text links with a
\href{https://en.wikipedia.org/wiki/Web_colors#Hex_triplet}{hexadecimal color value}.}
    \item{\code{widget_type}}{Set to \code{video} to return a Twitter Video embed for the
given Tweet.}
    \item{\code{dnt}}{When set to \code{true}, the Tweet and its embedded page on your site
are not used for purposes that include
\href{https://support.twitter.com/articles/20169421}{personalized suggestions}
and \href{https://support.twitter.com/articles/20170405}{personalized ads}.}
  }}
}
\description{
Takes a screenshot of the tweet that is automatically embedded in R Markdown
documents or that can be saved for use later.
}
\examples{
\dontrun{
tmpimg <- tempfile(fileext = "png")
tweet_screenshot(
  tweet_url = "https://twitter.com/alexpghayes/status/1211748406730706944",
  file = tmpimg
)
}

}
\seealso{
\code{\link[=tweet_embed]{tweet_embed()}}
}
